Fix MySQL Execute Scalar (#27)
* upgade angular cli 9

* update package lock

* Migrate to Angular 9, includes some long maintainace projects

* fix mysql execute scalar

+++ src/web-apis/LetPortal.Portal/Executions/MySQL/MySqlDynamicListQueryDatabase.cs
@@ -132,7 +132,7 @@ namespace LetPortal.Portal.Executions.MySQL
                             }
 
                         }
-                        response.TotalItems = (int)cmd.ExecuteScalar();
+                        response.TotalItems = (long)cmd.ExecuteScalar();
                     }
                 }
                 mysqlDbConnection.Close();
